Complete with  test _unittest_raft_leader_changes() (#11)
* Changed the logic of the terms, and also changed the tests

* Update tests/raft_node.py

Co-authored-by: maksimdrachov <39975120+maksimdrachov@users.noreply.github.com>

* Made some minor changes

* In node.py: expanded the log information for some values ​​to make it easier to analyze the code

In raft_log_replication.py: start to write test that log replication happens correctly if leadership changes

In raft_node.py: add a couple of lines of code to close node after each test, so that after running pytest it does not crash

new issue: after the leader changes, the new leader is not aware of the latest index of the neighboring nodes and begins to rewrite them, even though his nodes completely coincide with the neighboring node.

* Added test _unittest_raft_leader_changes for log_replication.py

* In node.py: expanded the log information for some values ​​to make it easier to analyze the code

In raft_log_replication.py: start to write test that log replication happens correctly if leadership changes

In raft_node.py: add a couple of lines of code to close node after each test, so that after running pytest it does not crash

new issue: after the leader changes, the new leader is not aware of the latest index of the neighboring nodes and begins to rewrite them, even though his nodes completely coincide with the neighboring node.

* Added test _unittest_raft_leader_changes for log_replication.py

* Adapted the test to term Changes

* Returned the deleted fragment from the test _unittest_raft_log_replication()

* Reformatted code

* Returned the test _unittest_raft_node_heartbeat()

* Added Black Formatter to the VSCode

* Formatted tests

* Changed raft_node.py tests

---------

Co-authored-by: maksimdrachov <39975120+maksimdrachov@users.noreply.github.com>